enforce lint layer contract: move scientific-notation-scan to malformed/
language/ must not import from rules/gml/ per the lint workspace layer
contract (malformed/ → language/ → rules/gml/).

recovery.ts (language/) was importing forEachScientificNotationToken from
rules/gml/scientific-notation-scan.ts, crossing the layer boundary upward.

scientific-notation-scan.ts is a pure text scanner (no AST required) that
operates on raw source characters to locate scientific-notation tokens.
This makes it a malformed/-tier utility, not a rule-layer helper.

- Move scientific-notation-scan.ts from rules/gml/ → malformed/
- Export the new symbols from malformed/index.ts
- Update language/recovery.ts to import from ../malformed/scientific-notation-scan.js
- Update no-scientific-notation-rule.ts to import from ../../../malformed/scientific-notation-scan.js
- Remove the old rules/gml/scientific-notation-scan.ts

All 547 lint tests pass. No TypeScript errors.
