+++ src/lint/src/malformed/scientific-notation-scan.ts
